+++ CHANGELOG.md
@@ -9,6 +9,19 @@ The format of this changelog is based on
 
 ## [Unreleased]
 
+- Update `npm_tags` to semi-match output with `docker_tags`.
+- Allow simple bash steps to be inserted into Github actions. This may be
+  necessary if we need to run `pip` commands or any other setup steps before the
+  main python action.
+- Fix: when using Github actions we need to specify the path relative to the
+  root of the project. In this case when we specify an action file we need to
+  find out the number of parent directories to be able to obtain the root.
+- _feat_: Add `m npm add_tags`. This command allows us to add tags to a package
+  in the npm registry. This is useful when we want to add tags to a package that
+  has already been published. Note that this will add docker tags generated
+  based on the published version.
+- Add next step instructions to `m start_release` and `m review_release`.
+
 ## [0.34.3] <a name="0.34.3" href="#0.34.3">-</a> May 23, 2024
 
 - Use `--load` with `docker buildx build`. See